Add files via upload
This script takes any tab delimited file, where the first column contains the known categories, and the rest of the columns contain any numeric data to be used as input into the Multi-Layer Perceptron.

An MLPClassifier is an implementation of a Multi-Layer Perceptron Classifier, which is a feedforward artificial neural network that maps the input dataset to a set of output classes.
                                                                                                                                                    
The confusion matrix helps you look at the errors in more detail. If we were just classifying True/False here, then the matrix would show correct answers on the diagonal (true positives and true negatives, and then false positives and false negatives on the other corners.                                                                                   
                                                                                                                                                     
The example I used to play with this script is a tab delimited file with tetranucleotided counts for 3201 MAGs (Meta-Genome Assembled Genomes), that have been classified into five known species of bacteria. In this case, the five column totals show the actual number of members for each species in the set, and the row totals show the species reported by the Perceptron. In this type of case, it could show whether particular pairs of species are getting mixed up more than others.                                                             
                                                                                                         
If you'd like to play with tetranucleotide statistics, please see get_kmer_frequencies.py, a script I wrote that takes a fasta file, calculates the tetranucleotides and writes them to a tab delimited file.